Add basic crash handler to handle exceptions
This is still an unpolished and bare-bone version, it will be improved
overtime, but for now it allows to see the various error messages in
case of a crash.

The crash still has to be saved to a file, display to the end user and
handle the various cases that may cause the crash handler to crash
(that's weird to type, but it can happen: I've seen it before).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>