Rework GenericLoader internal logic and public API
The internal caching of the resource allows to obtain references to
already loaded files (e.g. a dialog) without having to read them again
from disk and thus cut down on the I/O operations. At the same time,
this shares the controller object, but since they should be state-less
this shouldn't cause any problems. Controllers can always implement
a reset method to account for this (TODO: Add Resettable interface and
handle it in GenericLoader itself).

The rework of the API also allows for an easier-to-use external
interface and more unification in naming, that now resemble getters,
much more similar to their internal implementation. At the same time, it
is now possible to obtain the controller first, instead of having to
load the parent before working with it. A slight issue with that is that
now also the #getController() method may throw an IOException. Client
code will have to handle that situation too.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>