Add basic scheduler task
This addition allows the application to schedule tasks that have to be
run after a certain amount of time (e.g., a clock synchronizer).

For now this is merely a (thread-safe) wrapper around a
ScheduledExecutorService, but new features will be added as needed. In
other words, do not consider this API as stable because it is still in a
WIP stage.

TODO: allow tasks to schedule other tasks after a certain amount of time
from the last (probably custom Runnable class).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>