Create Worker Thread
The purpose of the worker thread is to execute the various tasks that
may or may not be run on the JavaFX thread or other threads without
interrupting the other operations.

This is to be preferred instead of the scheduler if the execution should
be performed immediately (or, better, as soon as possible, considering
it is still a single thread).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>